EE-1351 Fix comments
